+++ cardiosonix/pipeline/datasets/__init__.py
@@ -5,11 +5,10 @@ transform them and extract features
 using a third-party ETLPipeline module and load into RAM.
 
 Modules:
-    * classification.py - module contains a ``Dataset`` subclass named ``CardioAnomalyDataset``
+    * classify.py - module contains a ``Dataset`` subclass named ``CardioAnomalyDataset``
     for classifying heart rate deviations by sound.
 
-    * segmentation.py - at the moment, the module is not implemented and is not a priority.
+    * segment.py - at the moment, the module is not implemented and is not a priority.
 """
 
-from .classification import *
-
+from .classify import *
